% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_element_sequence}
\alias{build_element_sequence}
\title{Build the element value sequence for a pair of start and end value.}
\usage{
build_element_sequence(strt, end)
}
\arguments{
\item{strt}{Numeric start value of sequence}

\item{end}{Numeric end value of sequence}
}
\description{
Build the element value sequence for a pair of start and end value.
}
\keyword{internal}
