% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elements_to_vector}
\alias{elements_to_vector}
\title{Convert the elements columns in SWAT input format into vectors of individual
integer values}
\usage{
elements_to_vector(tbl)
}
\arguments{
\item{tbl}{SWAT+ input table which has the columns elem_tot and following
elements columns.}
}
\value{
The `tbl` where the columns `elem_tot` and `elements_*` are replaced
  by a list column with integer vectors providing all individual element IDs.
}
\description{
Convert the elements columns in SWAT input format into vectors of individual
integer values
}
