% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_num}
\alias{is_num}
\title{Identify whether all values of a vector are numeric or not}
\usage{
is_num(x)
}
\arguments{
\item{x}{Character vector}
}
\description{
Identify whether all values of a vector are numeric or not
}
\keyword{internal}
