% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swat.R
\name{lookup_read_type}
\alias{lookup_read_type}
\title{Look up the function type and additional input arguments to read a SWAT+ file.}
\usage{
lookup_read_type(file_name)
}
\arguments{
\item{file_name}{Path to the SWAT+ file.}
}
\value{
A list of arguments:
 - `type` is used to select the read function for the file.
 - `n_skip` and `has_unit` as arguments for the function `read_tbl`
 - or `def_names`, `par_names` and `id_num` as arguments for the
   function `read_tbl2`
 - `n_skip` as arguments for the function `read_tbl_n`
}
\description{
Look up the function type and additional input arguments to read a SWAT+ file.
}
\keyword{internal}
