% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_con_long}
\alias{pivot_con_long}
\title{Turn a .con input table into a long format table}
\usage{
pivot_con_long(con_tbl)
}
\arguments{
\item{con_tbl}{.con input table in (default) wide format}
}
\value{
The .con input table in a long format (each object connectivity
  will be a new row). Because of name duplicate the column name 'obj_id'
  of the defined unit is renamed to 'obj_id_'.
}
\description{
Turn a .con input table into a long format table
}
