% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_con_wide}
\alias{pivot_con_wide}
\title{Turn a .con input table in its long format back to its initial wide format.}
\usage{
pivot_con_wide(con_tbl)
}
\arguments{
\item{con_tbl}{.con input table in long format.}
}
\value{
The .con input table in a wide format (each object will be one
row, and all connectivities are in the same row). 'obj_id_' which is
renamed by `pivot_con_long` is renamed to 'obj_id' again.
}
\description{
Turn a .con input table in its long format back to its initial wide format.
}
