% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_con_file}
\alias{read_con_file}
\title{Read a SWAT+ connecitivity (*.con) input file.}
\usage{
read_con_file(file_path)
}
\arguments{
\item{file_path}{Path of the SWAT+ input file.}
}
\value{
The connecitivity input file as a tibble.
}
\description{
Read a SWAT+ connecitivity (*.con) input file.
}
\keyword{internal}
