% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_linewise}
\alias{read_linewise}
\title{Read the n column of a tabular SWAT+ input file which are defined by the
column positions `id_col_sel`. This is useful if e.g. last columns with
description cause issues with reading due to blanks in the description text.}
\usage{
read_linewise(file_path, col_names = NULL, n_skip = 1)
}
\arguments{
\item{file_path}{Path of the SWAT+ input file.}

\item{col_names}{(optional) Character column names vector.}

\item{n_skip}{Number of header rows to skip. Default is 1.}
}
\value{
The SWAT+ management.sch input file as a tibble.
}
\description{
Read the n column of a tabular SWAT+ input file which are defined by the
column positions `id_col_sel`. This is useful if e.g. last columns with
description cause issues with reading due to blanks in the description text.
}
