% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swat.R
\name{read_swat}
\alias{read_swat}
\title{Read SWAT+ input and output files.}
\usage{
read_swat(file_path)
}
\arguments{
\item{file_path}{Path to the SWAT+ file.}
}
\value{
The SWAT+ input or output file in a tibble.
}
\description{
Read SWAT+ input and output files.
}
