% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_tbl2}
\alias{read_tbl2}
\title{Read a SWAT+ input file which has a tabular structure with a definition line
for each parameter table section (e.g. management.sch, plant.ini, soils.sol).}
\usage{
read_tbl2(file_path, def_names, par_names, id_num = NULL)
}
\arguments{
\item{file_path}{Path of the SWAT+ input file.}

\item{def_names}{Vector of column names for the entries in the definition
line.}

\item{par_names}{Vector of parameter names of the parameter table.}

\item{id_num}{ID vector to define the columns which are numerical values.}
}
\value{
The SWAT+ management.sch input file as a tibble.
}
\description{
Read a SWAT+ input file which has a tabular structure with a definition line
for each parameter table section (e.g. management.sch, plant.ini, soils.sol).
}
\keyword{internal}
