% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_tbl_n}
\alias{read_tbl_n}
\title{Read the n column of a tabular SWAT+ input file which are defined by the
column positions `id_col_sel`. This is useful if e.g. last columns with
description cause issues with reading due to blanks in the description text.}
\usage{
read_tbl_n(
  file_path,
  col_names = NULL,
  n_skip = 1,
  id_col_sel = NULL,
  id_num = NULL
)
}
\arguments{
\item{file_path}{Path of the SWAT+ input file.}

\item{col_names}{(optional) Character column names vector.}

\item{n_skip}{Number of header rows to skip. Default is 1.}

\item{id_col_sel}{Numeric vector which defines the column positions that are
returned in the table.}

\item{id_num}{ID vector to define the columns which are numerical values.}
}
\value{
The SWAT+ management.sch input file as a tibble.
}
\description{
Read the n column of a tabular SWAT+ input file which are defined by the
column positions `id_col_sel`. This is useful if e.g. last columns with
description cause issues with reading due to blanks in the description text.
}
\keyword{internal}
