% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_elements}
\alias{sum_elements}
\title{Sum the elements which are not NA}
\usage{
sum_elements(elem)
}
\arguments{
\item{elem}{Element entries.}
}
\description{
Sum the elements which are not NA
}
\keyword{internal}
