% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{values_to_elements}
\alias{values_to_elements}
\title{Convert the information on available runs for the simulated variables into
strings that are printed}
\usage{
values_to_elements(vals)
}
\arguments{
\item{tbl}{overview table that provides meta data for all simulation runs for
all variables saved in the data bases}
}
\description{
Convert the information on available runs for the simulated variables into
strings that are printed
}
\keyword{internal}
