% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector_to_elements}
\alias{vector_to_elements}
\title{Convert the elements list column into the columns elem_tot and elements_*
which are the default SWAT+ format.
integer values}
\usage{
vector_to_elements(tbl)
}
\arguments{
\item{tbl}{SWAT+ input table which has the list column elements.}
}
\value{
The `tbl` where the list column `elements` is converted into the
  columns `elem_tot` and `elements_*`.
}
\description{
Convert the elements list column into the columns elem_tot and elements_*
which are the default SWAT+ format.
integer values
}
