% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_swat.R
\name{write_swat}
\alias{write_swat}
\title{Write SWAT+ input files.}
\usage{
write_swat(tbl, file_path, overwrite = FALSE)
}
\arguments{
\item{tbl}{SWAT input table}

\item{file_path}{Write path of the SWAT+ file.}

\item{overwrite}{Should original file be overwritten or renamed before writing?}
}
\value{
Writes a text file table in the file path.
}
\description{
Write SWAT+ input files.
}
