% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_tbl2}
\alias{write_tbl2}
\title{Write SWAT+ input file which has a tabular structure with a definition line
for each parameter table section (e.g. management.sch, plant.ini, soils.sol).}
\usage{
write_tbl2(tbl, file_path, fmt_def, fmt_par)
}
\arguments{
\item{tbl}{SWAT input table in tibble (data.frame) format.}

\item{file_path}{Write path of the SWAT+ input file.}

\item{fmt_def}{Character vector of format strings to define the print format
of each table column which is part of the definition line.}

\item{fmt_par}{Character vector of format strings to define the print format
of each table column which is part of the parameter table.}
}
\value{
Writes a text file table in the file path.
}
\description{
Write SWAT+ input file which has a tabular structure with a definition line
for each parameter table section (e.g. management.sch, plant.ini, soils.sol).
}
\keyword{internal}
